reduce code duplication in tests
