+++ test/test_helper.exs
@@ -136,3 +136,62 @@ defmodule Pulsar.TestHelper do
     Process.sleep(1000)
   end
 end
+
+# Shared test callback module for all integration tests
+defmodule Pulsar.DummyConsumer do
+  @moduledoc """
+  A test consumer callback that collects messages and provides test utilities.
+
+  This module can be used across different test files to avoid code duplication.
+  """
+
+  @behaviour Pulsar.Consumer.Callback
+
+  def init(_opts) do
+    {:ok, %{messages: [], count: 0}}
+  end
+
+  def handle_message(message, state) do
+    new_state = %{
+      state
+      | messages: [message | state.messages],
+        count: state.count + 1
+    }
+
+    {:ok, new_state}
+  end
+
+  # Custom GenServer calls for testing
+  def handle_call(:get_messages, _from, state) do
+    {:reply, Enum.reverse(state.messages), state}
+  end
+
+  def handle_call(:count_messages, _from, state) do
+    {:reply, state.count, state}
+  end
+
+  def handle_call(:get_state, _from, state) do
+    {:reply, state, state}
+  end
+
+  def handle_cast(:clear_messages, _state) do
+    {:noreply, %{messages: [], count: 0}}
+  end
+
+  # Test helper functions that delegate to GenServer calls
+  def get_messages(consumer_pid) do
+    GenServer.call(consumer_pid, :get_messages)
+  end
+
+  def clear_messages(consumer_pid) do
+    GenServer.cast(consumer_pid, :clear_messages)
+  end
+
+  def count_messages(consumer_pid) do
+    GenServer.call(consumer_pid, :count_messages)
+  end
+
+  def get_state(consumer_pid) do
+    GenServer.call(consumer_pid, :get_state)
+  end
+end
